+++ extensions/vscode/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.0.3 (06/16/26)
+
+- Misc. bug fixes with lists, tasklist items, and paste
+
 ## v0.0.2 (06/15/26)
 
 - Added basic syntax highlighting support for Markdown and JavaScript/TypeScript